editformepls

+++ src/main/java/com/Boardzone/boardapi/services/BoardGameService.java
@@ -4,10 +4,10 @@ import com.Boardzone.boardapi.entity.BoardGameEntity;
 import java.util.List;
 
 public interface BoardGameService {
-    BoardGameEntity addBoardGame(String boardgameName, int maxPlayers, long lobbyId, byte[] imagesetId, String description);
+    BoardGameEntity addBoardGame(String boardgameName, int maxPlayers, int lobbyId, byte[] imagesetId, String description);
     List<BoardGameEntity> getAllBoardGames();
-    BoardGameEntity getBoardGame(long boardgameId);
-    void deleteBoardGame(long boardgameId);
-    BoardGameEntity updateBoardGame(long boardgameId, String boardgameName, int maxPlayers, long lobbyId, byte[] imagesetId, String description);
-    long findFirstAvailableId();
+    BoardGameEntity getBoardGame(int boardgameId); 
+    void deleteBoardGame(int boardgameId); 
+    BoardGameEntity updateBoardGame(int boardgameId, String boardgameName, int maxPlayers, int lobbyId, byte[] imagesetId, String description); 
+    int findFirstAvailableId(); 
 }
